refactor: consolidate vc meeting actions
